migration fix

+++ database/migrations/2019_04_22_150658_create_reviews_table.php
@@ -17,7 +17,7 @@ class CreateReviewsTable extends Migration
             $table->bigIncrements('id');
             $table->timestamps();
             $table->string('review');
-            $table->integer('rate');
+            $table->float('rate');
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('book_id');
             $table->foreign('user_id')->on('users')->references('id');
